+++ CotEditor/Sources/CEPrintView.m
@@ -36,6 +36,7 @@
 
 #import "NSString+Sandboxing.h"
 #import "NSString+CECounting.h"
+#import "NSFont+CESize.h"
 
 
 // constants
@@ -284,10 +285,9 @@ static NSString *_Nonnull const PageNumberPlaceholder = @"PAGENUM";
     // set tab width
     NSMutableParagraphStyle *paragraphStyle = [[NSParagraphStyle defaultParagraphStyle] mutableCopy];
     NSUInteger tabWidth = [[NSUserDefaults standardUserDefaults] integerForKey:CEDefaultTabWidthKey];
-    CGFloat spaceWidth = [font advancementForGlyph:(NSGlyph)' '].width;
     
     [paragraphStyle setTabStops:@[]];
-    [paragraphStyle setDefaultTabInterval:tabWidth * spaceWidth];
+    [paragraphStyle setDefaultTabInterval:tabWidth * [font advancementForCharacter:' ']];
     [self setDefaultParagraphStyle:paragraphStyle];
     
     // apply to current string